supply paths via cmd line

+++ classification/distilbert_baseline.py
@@ -104,7 +104,7 @@ def train(train_dataset, model, train_params, class_weights=None):
     )
 
     if class_weights is not None:
-        class_weights = torch.from_numpy(class_weights).to(device)
+        class_weights = torch.from_numpy(class_weights).float().to(device)
 
     no_decay = {'bias', 'LayerNorm.weight'}
     optimizer_grouped_parameters = [
